+++ src/ExpertSenderApi.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 namespace Citilink\ExpertSenderApi;
 
 use Citilink\ExpertSenderApi\Resource\BouncesResource;
+use Citilink\ExpertSenderApi\Resource\DataTablesResource;
+use Citilink\ExpertSenderApi\Resource\MessagesResource;
 use Citilink\ExpertSenderApi\Resource\RemovedSubscribersResource;
 use Citilink\ExpertSenderApi\Resource\SubscribersResource;
 use Citilink\ExpertSenderApi\Resource\TimeResource;
@@ -79,5 +81,25 @@ class ExpertSenderApi
     public function removedSubscribers(): RemovedSubscribersResource
     {
         return new RemovedSubscribersResource($this->requestSender);
+    }
+
+    /**
+     * Get data tables resource
+     *
+     * @return DataTablesResource Data tables resource
+     */
+    public function dataTables(): DataTablesResource
+    {
+        return new DataTablesResource($this->requestSender);
+    }
+
+    /**
+     * Get messages resource
+     *
+     * @return MessagesResource Messages resource
+     */
+    public function messages(): MessagesResource
+    {
+        return new MessagesResource($this->requestSender);
     }
 }