+++ src/Model/DataTablesAddMultipleRowsPostRequest/Row.php
@@ -1,0 +1,42 @@
+<?php
+declare(strict_types=1);
+
+namespace Citilink\ExpertSenderApi\Model\DataTablesAddMultipleRowsPostRequest;
+
+use Citilink\ExpertSenderApi\Model\Column;
+use Webmozart\Assert\Assert;
+
+/**
+ * Row
+ *
+ * @author Nikita Sapogov <sapogov.n@citilink.ru>
+ */
+class Row
+{
+    /**
+     * @var Column[]|iterable Columns
+     */
+    private $columns;
+
+    /**
+     * Constructor
+     *
+     * @param Column[]|iterable $columns Columns
+     */
+    public function __construct(iterable $columns)
+    {
+        Assert::notEmpty($columns);
+        $this->columns = $columns;
+    }
+
+    /**
+     * Get columns
+     *
+     * @return Column[]|iterable Columns
+     */
+    public function getColumns()
+    {
+        return $this->columns;
+    }
+
+}